refactor: move internal to pkg
